move dbConnect.js from /lib to /db
