+++ README.md
@@ -2,18 +2,36 @@
 
 Tech Stack:
 
+Frontend:
 - NextJS / React
 - TailwindCSS
+
+Backend
+- ExpressJS
 - PostGreSQL
 - NodeJS
 
 ## Getting Started
 
-First, run `./setup.sh`
+### Frontend
+
+First, run `./setup.sh` or `npm run setup`
 Then run the development server:
 `npm run dev`
+To build and run regular server:
+`npm run build && npm run start`
 
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+
+### Backend
+
+Inside the /server/ folder:
+Development server (nodemon reloading):
+`npm run dev`
+Regular server
+`npm run start`
+
+API server lives on [http://localhost:3001/api](http://localhost:3001/api)
 
 ## Contributing
 
@@ -23,11 +41,9 @@ Or create PR or Fork of this project.
 
 ### Learn More
 
-The `pages/api` directory is mapped to `http://localhost:3000/api/*`. Files in this directory are treated as [API routes](https://nextjs.org/docs/api-routes/introduction) instead of React pages.
-
 To learn more about Next.js, take a look at the following resources:
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
+- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features
 - [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
 
 You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
