+++ server/controllers/comment/deleteComment.js
@@ -1,4 +1,4 @@
-const pool = require("../../lib/dbConnect.js");
+const pool = require("../../db/dbConnect.js");
 
 //Delete a post
 async function deleteComment(commentID) {
